+++ test/commands/kidney_do_test.dart
@@ -44,6 +44,7 @@ void main() {
           'publish',
           'review',
           'install-git-hooks',
+          'cancel-review',
         ],
       );